Minor code enhances.

+++ test/homework/ParserTest.java
@@ -4,7 +4,6 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.PrintWriter;
 
@@ -23,16 +22,18 @@ public class ParserTest {
       w.println("0;1;2;3;4");
       for (String[] line : Parser.parse(f)) {
         for (int i = 0; i < 5; i++) {
-          assertEquals(i, line[i]);
+          assertEquals("" + i, line[i]);
         }
       }
-    } catch (FileNotFoundException e) {
-      fail();
     } catch (IOException e) {
       fail();
     } finally {
-      w.close();
-      f.delete();
+      if (w != null) {
+        w.close();
+      }
+      if (f != null) {
+        f.delete();
+      }
     }
   }
 }
